move base model to utils
