+++ runway/util.py
@@ -37,6 +37,7 @@ from typing import (
 )
 
 import yaml
+from pydantic import BaseModel as _BaseModel
 from typing_extensions import Literal
 
 # make this importable for util as it was before
@@ -48,6 +49,54 @@ if TYPE_CHECKING:
 AWS_ENV_VARS = ("AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY", "AWS_SESSION_TOKEN")
 DOC_SITE = "https://docs.onica.com/projects/runway"
 EMBEDDED_LIB_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), "embedded")
+
+
+class BaseModel(_BaseModel):
+    """Base class for Runway models."""
+
+    def get(self, name: str, default: Any = None) -> Any:
+        """Implement evaluation of self.get.
+
+        Args:
+            name: Attribute name to return the value for.
+            default: Value to return if attribute is not found.
+
+        """
+        return getattr(self, name, default)
+
+    def __contains__(self, name: str) -> bool:
+        """Implement evaluation of 'in' conditional.
+
+        Args:
+            name: The name to check for existence in the model.
+
+        """
+        return name in self.__dict__
+
+    def __getitem__(self, name: str) -> Any:
+        """Implement evaluation of self[name].
+
+        Args:
+            name: Attribute name to return the value for.
+
+        Returns:
+            The value associated with the provided name/attribute name.
+
+        Raises:
+            AttributeError: If attribute does not exist on this object.
+
+        """
+        return getattr(self, name)
+
+    def __setitem__(self, name: str, value: Any) -> None:
+        """Implement item assignment (e.g. ``self[name] = value``).
+
+        Args:
+            name: Attribute name to set.
+            value: Value to assign to the attribute.
+
+        """
+        super().__setattr__(name, value)  # type: ignore
 
 
 class JsonEncoder(json.JSONEncoder):
@@ -459,6 +508,15 @@ def ensure_file_is_executable(path: str) -> None:
     ):
         print("Error: File %s is not executable" % path)
         sys.exit(1)
+
+
+def ensure_string(value: Any) -> str:
+    """Ensure value is a string."""
+    if isinstance(value, str):
+        return value
+    if isinstance(value, bytes):
+        return value.decode()
+    raise TypeError(f"Expected str or bytes but received {type(value)}")
 
 
 @contextmanager
